+++ yfinance/base.py
@@ -22,6 +22,7 @@
 from __future__ import print_function
 
 import datetime as _datetime
+from io import StringIO
 import json as _json
 import logging
 import time as _time
@@ -34,9 +35,8 @@ import numpy as np
 import pandas as pd
 import requests
 
-from . import shared
-from . import utils
-from .data import TickerData
+from . import shared, utils, cache
+from .data import YfData
 from .scrapers.analysis import Analysis
 from .scrapers.fundamentals import Fundamentals
 from .scrapers.holders import Holders
@@ -72,12 +72,12 @@ class TickerBase:
         if utils.is_isin(self.ticker):
             self.ticker = utils.get_ticker_by_isin(self.ticker, None, session)
 
-        self._data: TickerData = TickerData(self.ticker, session=session)
+        self._data: YfData = YfData(session=session)
 
-        self._analysis = Analysis(self._data)
-        self._holders = Holders(self._data)
-        self._quote = Quote(self._data)
-        self._fundamentals = Fundamentals(self._data)
+        self._analysis = Analysis(self._data, self.ticker)
+        self._holders = Holders(self._data, self.ticker)
+        self._quote = Quote(self._data, self.ticker)
+        self._fundamentals = Fundamentals(self._data, self.ticker)
 
         self._fast_info = None
 
@@ -89,7 +89,6 @@ class TickerBase:
                 start=None, end=None, prepost=False, actions=True,
                 auto_adjust=True, back_adjust=False, repair=False, keepna=False,
                 proxy=None, rounding=False, timeout=10,
-                debug=None,  # deprecated
                 raise_errors=False) -> pd.DataFrame:
         """
         :Parameters:
@@ -129,22 +128,11 @@ class TickerBase:
                 If not None stops waiting for a response after given number of
                 seconds. (Can also be a fraction of a second e.g. 0.01)
                 Default is 10 seconds.
-            debug: bool
-                If passed as False, will suppress message printing to console.
-                DEPRECATED, will be removed in future version
             raise_errors: bool
                 If True, then raise errors as Exceptions instead of logging.
         """
         logger = utils.get_yf_logger()
         proxy = proxy or self.proxy
-
-        if debug is not None:
-            if debug:
-                utils.print_once(f"yfinance: Ticker.history(debug={debug}) argument is deprecated and will be removed in future version. Do this instead: logging.getLogger('yfinance').setLevel(logging.ERROR)")
-                logger.setLevel(logging.ERROR)
-            else:
-                utils.print_once(f"yfinance: Ticker.history(debug={debug}) argument is deprecated and will be removed in future version. Do this instead to suppress error messages: logging.getLogger('yfinance').setLevel(logging.CRITICAL)")
-                logger.setLevel(logging.CRITICAL)
 
         start_user = start
         end_user = end
@@ -221,7 +209,8 @@ class TickerBase:
 
             data = data.json()
         except Exception:
-            pass
+            if raise_errors:
+                raise
 
         # Store the meta data that gets retrieved simultaneously
         try:
@@ -250,9 +239,9 @@ class TickerBase:
             err_msg += f' (period={period})'
 
         fail = False
-        if data is None or type(data) is not dict:
+        if data is None or not isinstance(data, dict):
             fail = True
-        elif type(data) is dict and 'status_code' in data:
+        elif isinstance(data, dict) and 'status_code' in data:
             err_msg += f"(Yahoo status_code = {data['status_code']})"
             fail = True
         elif "chart" in data and data["chart"]["error"]:
@@ -397,9 +386,6 @@ class TickerBase:
 
         df = df[~df.index.duplicated(keep='first')]  # must do before repair
 
-        if isinstance(repair, str) and repair=='silent':
-            utils.log_once(logging.WARNING, f"yfinance: Ticker.history(repair='silent') value is deprecated and will be removed in future version. Repair now silent by default, use logging module to increase verbosity.")
-            repair = True
         if repair:
             # Do this before auto/back adjust
             logger.debug(f'{self.ticker}: checking OHLC for repairs ...')
@@ -524,7 +510,6 @@ class TickerBase:
                 return df
 
         dts_to_repair = df.index[f_repair_rows]
-        indices_to_repair = np.where(f_repair_rows)[0]
 
         if len(dts_to_repair) == 0:
             logger.info("Nothing needs repairing (dts_to_repair[] empty)")
@@ -541,9 +526,6 @@ class TickerBase:
 
         # Group nearby NaN-intervals together to reduce number of Yahoo fetches
         dts_groups = [[dts_to_repair[0]]]
-        last_dt = dts_to_repair[0]
-        last_ind = indices_to_repair[0]
-        td = utils._interval_to_timedelta(interval)
         # Note on setting max size: have to allow space for adding good data
         if sub_interval == "1mo":
             grp_max_size = _dateutil.relativedelta.relativedelta(years=2)
@@ -559,14 +541,11 @@ class TickerBase:
             grp_max_size = _datetime.timedelta(days=30)
         logger.debug(f"grp_max_size = {grp_max_size}")
         for i in range(1, len(dts_to_repair)):
-            ind = indices_to_repair[i]
             dt = dts_to_repair[i]
             if dt.date() < dts_groups[-1][0].date() + grp_max_size:
                 dts_groups[-1].append(dt)
             else:
                 dts_groups.append([dt])
-            last_dt = dt
-            last_ind = ind
 
         logger.debug("Repair groups:")
         for g in dts_groups:
@@ -690,7 +669,7 @@ class TickerBase:
                 df_new.index = new_index
             logger.debug('df_new:' + '\n' + str(df_new))
 
-            # Calibrate! 
+            # Calibrate!
             common_index = np.intersect1d(df_block.index, df_new.index)
             if len(common_index) == 0:
                 # Can't calibrate so don't attempt repair
@@ -704,38 +683,42 @@ class TickerBase:
                 f_tag = df_block_calib['Adj Close'] == tag
                 if f_tag.any():
                     div_adjusts = df_block_calib['Adj Close'] / df_block_calib['Close']
-                    # The loop below assumes each 1d repair is isoloated, i.e. surrounded by 
-                    # good data. Which is case most of time. 
-                    # But in case are repairing a chunk of bad 1d data, back/forward-fill the 
+                    # The loop below assumes each 1d repair is isolated, i.e. surrounded by
+                    # good data. Which is case most of time.
+                    # But in case are repairing a chunk of bad 1d data, back/forward-fill the
                     # good div-adjustments - not perfect, but a good backup.
                     div_adjusts[f_tag] = np.nan
-                    div_adjusts = div_adjusts.fillna(method='bfill').fillna(method='ffill')
+                    div_adjusts = div_adjusts.ffill().bfill()
                     for idx in np.where(f_tag)[0]:
                         dt = df_new_calib.index[idx]
                         n = len(div_adjusts)
                         if df_new.loc[dt, "Dividends"] != 0:
                             if idx < n - 1:
                                 # Easy, take div-adjustment from next-day
-                                div_adjusts[idx] = div_adjusts[idx + 1]
+                                div_adjusts.iloc[idx] = div_adjusts.iloc[idx + 1]
                             else:
                                 # Take previous-day div-adjustment and reverse todays adjustment
                                 div_adj = 1.0 - df_new_calib["Dividends"].iloc[idx] / df_new_calib['Close'].iloc[
                                     idx - 1]
-                                div_adjusts[idx] = div_adjusts[idx - 1] / div_adj
+                                div_adjusts.iloc[idx] = div_adjusts.iloc[idx - 1] / div_adj
                         else:
                             if idx > 0:
                                 # Easy, take div-adjustment from previous-day
-                                div_adjusts[idx] = div_adjusts[idx - 1]
+                                div_adjusts.iloc[idx] = div_adjusts.iloc[idx - 1]
                             else:
                                 # Must take next-day div-adjustment
-                                div_adjusts[idx] = div_adjusts[idx + 1]
+                                div_adjusts.iloc[idx] = div_adjusts.iloc[idx + 1]
                                 if df_new_calib["Dividends"].iloc[idx + 1] != 0:
-                                    div_adjusts[idx] *= 1.0 - df_new_calib["Dividends"].iloc[idx + 1] / \
+                                    div_adjusts.iloc[idx] *= 1.0 - df_new_calib["Dividends"].iloc[idx + 1] / \
                                                         df_new_calib['Close'].iloc[idx]
                     f_close_bad = df_block_calib['Close'] == tag
+                    div_adjusts = div_adjusts.reindex(df_block.index, fill_value=np.nan).ffill().bfill()
                     df_new['Adj Close'] = df_block['Close'] * div_adjusts
                     if f_close_bad.any():
-                        df_new.loc[f_close_bad, 'Adj Close'] = df_new['Close'][f_close_bad] * div_adjusts[f_close_bad]
+                        f_close_bad_new = f_close_bad.reindex(df_new.index, fill_value=False)
+                        div_adjusts_new = div_adjusts.reindex(df_new.index, fill_value=np.nan).ffill().bfill()
+                        div_adjusts_new_np = f_close_bad_new.to_numpy()
+                        df_new.loc[div_adjusts_new_np, 'Adj Close'] = df_new['Close'][div_adjusts_new_np] * div_adjusts_new[div_adjusts_new_np]
 
             # Check whether 'df_fine' has different split-adjustment.
             # If different, then adjust to match 'df'
@@ -790,14 +773,14 @@ class TickerBase:
             no_fine_data_dts = []
             for idx in bad_dts:
                 if idx not in df_new.index:
-                    # Yahoo didn't return finer-grain data for this interval, 
+                    # Yahoo didn't return finer-grain data for this interval,
                     # so probably no trading happened.
                     no_fine_data_dts.append(idx)
             if len(no_fine_data_dts) > 0:
-                logger.debug(f"Yahoo didn't return finer-grain data for these intervals: " + str(no_fine_data_dts))
+                logger.debug("Yahoo didn't return finer-grain data for these intervals: " + str(no_fine_data_dts))
             for idx in bad_dts:
                 if idx not in df_new.index:
-                    # Yahoo didn't return finer-grain data for this interval, 
+                    # Yahoo didn't return finer-grain data for this interval,
                     # so probably no trading happened.
                     continue
                 df_new_row = df_new.loc[idx]
@@ -932,11 +915,13 @@ class TickerBase:
                     if fi[j]:
                         df2.loc[idx, c] = df.loc[idx, c] * 0.01
 
-                c = "High" ; j = data_cols.index(c)
+                c = "High"
+                j = data_cols.index(c)
                 if fi[j]:
                     df2.loc[idx, c] = df2.loc[idx, ["Open", "Close"]].max()
 
-                c = "Low" ; j = data_cols.index(c)
+                c = "Low"
+                j = data_cols.index(c)
                 if fi[j]:
                     df2.loc[idx, c] = df2.loc[idx, ["Open", "Close"]].min()
 
@@ -952,11 +937,13 @@ class TickerBase:
                     if fi[j]:
                         df2.loc[idx, c] = df.loc[idx, c] * 100.0
 
-                c = "High" ; j = data_cols.index(c)
+                c = "High"
+                j = data_cols.index(c)
                 if fi[j]:
                     df2.loc[idx, c] = df2.loc[idx, ["Open", "Close"]].max()
 
-                c = "Low" ; j = data_cols.index(c)
+                c = "Low"
+                j = data_cols.index(c)
                 if fi[j]:
                     df2.loc[idx, c] = df2.loc[idx, ["Open", "Close"]].min()
 
@@ -1145,7 +1132,7 @@ class TickerBase:
         div_indices = np.where(f_div)[0]
         last_div_idx = div_indices[-1]
         if last_div_idx == 0:
-            # Not enough data to recalculate the div-adjustment, 
+            # Not enough data to recalculate the div-adjustment,
             # because need close day before
             logger.debug('div-adjust-repair: Insufficient data to recalculate div-adjustment')
             return df
@@ -1189,13 +1176,13 @@ class TickerBase:
         # stock split to old price data.
         #
         # There is a slight complication, because Yahoo does another stupid thing.
-        # Sometimes the old data is adjusted twice. So cannot simply assume 
-        # which direction to reverse adjustment - have to analyse prices and detect. 
+        # Sometimes the old data is adjusted twice. So cannot simply assume
+        # which direction to reverse adjustment - have to analyse prices and detect.
         # Not difficult.
 
         if df.empty:
             return df
-            
+
         logger = utils.get_yf_logger()
 
         interday = interval in ['1d', '1wk', '1mo', '3mo']
@@ -1223,7 +1210,7 @@ class TickerBase:
     def _fix_prices_sudden_change(self, df, interval, tz_exchange, change, correct_volume=False):
         if df.empty:
             return df
-            
+
         logger = utils.get_yf_logger()
 
         df = df.sort_index(ascending=False)
@@ -1242,9 +1229,8 @@ class TickerBase:
             logger.debug(f'price-repair-split: start_min={start_min}')
 
         OHLC = ['Open', 'High', 'Low', 'Close']
-        OHLCA = OHLC + ['Adj Close']
 
-        # Do not attempt repair of the split is small, 
+        # Do not attempt repair of the split is small,
         # could be mistaken for normal price variance
         if 0.8 < split < 1.25:
             logger.info("price-repair-split: Split ratio too close to 1. Won't repair")
@@ -1282,10 +1268,10 @@ class TickerBase:
         else:
             debug_cols = []
 
-        # Calculate daily price % change. To reduce effect of price volatility, 
+        # Calculate daily price % change. To reduce effect of price volatility,
         # calculate change for each OHLC column.
         if interday and interval != '1d' and split not in [100.0, 100, 0.001]:
-            # Avoid using 'Low' and 'High'. For multiday intervals, these can be 
+            # Avoid using 'Low' and 'High'. For multiday intervals, these can be
             # very volatile so reduce ability to detect genuine stock split errors
             _1d_change_x = np.full((n, 2), 1.0)
             price_data = df2[['Open','Close']].to_numpy()
@@ -1300,8 +1286,13 @@ class TickerBase:
         # Update: if a VERY large dividend is paid out, then can be mistaken for a 1:2 stock split.
         # Fix = use adjusted prices
         adj = df2['Adj Close'].to_numpy() / df2['Close'].to_numpy()
+        df_dtype = price_data.dtype
+        if df_dtype == np.int64:
+            price_data = price_data.astype('float')
         for j in range(price_data.shape[1]):
             price_data[:,j] *= adj
+        if df_dtype == np.int64:
+            price_data = price_data.astype('int')
 
         _1d_change_x[1:] = price_data[1:, ] / price_data[:-1, ]
         f_zero_num_denom = f_zero | np.roll(f_zero, 1, axis=0)
@@ -1425,7 +1416,7 @@ class TickerBase:
                 else:
                     threshold = _datetime.timedelta(days=threshold_days)
                 if gap_td < threshold:
-                    logger.info(f'price-repair-split: 100x changes are too soon after stock split events, aborting')
+                    logger.info('price-repair-split: 100x changes are too soon after stock split events, aborting')
                     return df
 
         # if logger.isEnabledFor(logging.DEBUG):
@@ -1437,9 +1428,12 @@ class TickerBase:
         def map_signals_to_ranges(f, f_up, f_down):
             # Ensure 0th element is False, because True is nonsense
             if f[0]:
-                f = np.copy(f) ; f[0] = False
-                f_up = np.copy(f_up) ; f_up[0] = False
-                f_down = np.copy(f_down) ; f_down[0] = False
+                f = np.copy(f)
+                f[0] = False
+                f_up = np.copy(f_up)
+                f_up[0] = False
+                f_down = np.copy(f_down)
+                f_down[0] = False
 
             if not f.any():
                 return []
@@ -1470,9 +1464,9 @@ class TickerBase:
         if correct_columns_individually:
             f_corrected = np.full(n, False)
             if correct_volume:
-                # If Open or Close is repaired but not both, 
+                # If Open or Close is repaired but not both,
                 # then this means the interval has a mix of correct
-                # and errors. A problem for correcting Volume, 
+                # and errors. A problem for correcting Volume,
                 # so use a heuristic:
                 # - if both Open & Close were Nx bad => Volume is Nx bad
                 # - if only one of Open & Close are Nx bad => Volume is 0.5*Nx bad
@@ -1486,7 +1480,7 @@ class TickerBase:
                 if appears_suspended and (idx_latest_active is not None and idx_latest_active >= idx_first_f):
                     # Suspended midway during data date range.
                     # 1: process data before suspension in index-ascending (date-descending) order.
-                    # 2: process data after suspension in index-descending order. Requires signals to be reversed, 
+                    # 2: process data after suspension in index-descending order. Requires signals to be reversed,
                     #    then returned ranges to also be reversed, because this logic was originally written for
                     #    index-ascending (date-descending) order.
                     fj = f[:, j]
@@ -1511,7 +1505,8 @@ class TickerBase:
                         for i in range(len(ranges_after)):
                             r = ranges_after[i]
                             ranges_after[i] = (n-r[1], n-r[0], r[2])
-                    ranges = ranges_before ; ranges.extend(ranges_after)
+                    ranges = ranges_before
+                    ranges.extend(ranges_after)
                 else:
                     ranges = map_signals_to_ranges(f[:, j], f_up[:, j], f_down[:, j])
                 logger.debug(f"column '{c}' ranges: {ranges}")
@@ -1544,9 +1539,11 @@ class TickerBase:
                         ranges = []
                     for r in ranges:
                         if r[2] == 'split':
-                            m = split ; m_rcp = split_rcp
+                            m = split
+                            m_rcp = split_rcp
                         else:
-                            m = split_rcp ; m_rcp = split
+                            m = split_rcp
+                            m_rcp = split
                         if interday:
                             logger.info(f"price-repair-split: Corrected {fix_type} on col={c} range=[{df2.index[r[1]-1].date()}:{df2.index[r[0]].date()}] m={m:.4f}")
                         else:
@@ -1576,7 +1573,7 @@ class TickerBase:
             if appears_suspended and (idx_latest_active is not None and idx_latest_active >= idx_first_f):
                 # Suspended midway during data date range.
                 # 1: process data before suspension in index-ascending (date-descending) order.
-                # 2: process data after suspension in index-descending order. Requires signals to be reversed, 
+                # 2: process data after suspension in index-descending order. Requires signals to be reversed,
                 #    then returned ranges to also be reversed, because this logic was originally written for
                 #    index-ascending (date-descending) order.
                 ranges_before = map_signals_to_ranges(f[idx_latest_active:], f_up[idx_latest_active:], f_down[idx_latest_active:])
@@ -1598,7 +1595,8 @@ class TickerBase:
                     for i in range(len(ranges_after)):
                         r = ranges_after[i]
                         ranges_after[i] = (n-r[1], n-r[0], r[2])
-                ranges = ranges_before ; ranges.extend(ranges_after)
+                ranges = ranges_before
+                ranges.extend(ranges_after)
             else:
                 ranges = map_signals_to_ranges(f, f_up, f_down)
             if start_min is not None:
@@ -1610,9 +1608,11 @@ class TickerBase:
                         del ranges[i]
             for r in ranges:
                 if r[2] == 'split':
-                    m = split ; m_rcp = split_rcp
+                    m = split
+                    m_rcp = split_rcp
                 else:
-                    m = split_rcp ; m_rcp = split
+                    m = split_rcp
+                    m_rcp = split
                 logger.debug(f"price-repair-split: range={r} m={m}")
                 for c in ['Open', 'High', 'Low', 'Close', 'Adj Close']:
                     df2.iloc[r[0]:r[1], df2.columns.get_loc(c)] *= m
@@ -1643,16 +1643,16 @@ class TickerBase:
 
         return df2
 
-    def _get_ticker_tz(self,timeout, proxy=None):
+    def _get_ticker_tz(self, proxy, timeout):
         proxy = proxy or self.proxy
         if self._tz is not None:
             return self._tz
-        cache = utils.get_tz_cache()
-        tz = cache.lookup(self.ticker)
+        c = cache.get_tz_cache()
+        tz = c.lookup(self.ticker)
 
         if tz and not utils.is_valid_timezone(tz):
             # Clear from cache and force re-fetch
-            cache.store(self.ticker, None)
+            c.store(self.ticker, None)
             tz = None
 
         if tz is None:
@@ -1660,7 +1660,7 @@ class TickerBase:
 
             if utils.is_valid_timezone(tz):
                 # info fetch is relatively slow so cache timezone
-                cache.store(self.ticker, tz)
+                c.store(self.ticker, tz)
             else:
                 tz = None
 
@@ -1668,7 +1668,7 @@ class TickerBase:
         return tz
 
     @utils.log_indent_decorator
-    def _fetch_ticker_tz(self, timeout, proxy=None):
+    def _fetch_ticker_tz(self, proxy, timeout):
         # Query Yahoo for fast price data just to get returned timezone
         proxy = proxy or self.proxy
         logger = utils.get_yf_logger()
@@ -1701,18 +1701,34 @@ class TickerBase:
         return None
 
     def get_recommendations(self, proxy=None, as_dict=False):
+        """
+        Returns a DataFrame with the recommendations
+        Columns: period  strongBuy  buy  hold  sell  strongSell
+        """
         self._quote.proxy = proxy or self.proxy
         data = self._quote.recommendations
         if as_dict:
             return data.to_dict()
         return data
 
-    def get_calendar(self, proxy=None, as_dict=False):
+    def get_recommendations_summary(self, proxy=None, as_dict=False):
+        return self.get_recommendations(proxy=proxy, as_dict=as_dict)
+
+    def get_upgrades_downgrades(self, proxy=None, as_dict=False):
+        """
+        Returns a DataFrame with the recommendations changes (upgrades/downgrades)
+        Index: date of grade
+        Columns: firm toGrade fromGrade action
+        """
         self._quote.proxy = proxy or self.proxy
-        data = self._quote.calendar
+        data = self._quote.upgrades_downgrades
         if as_dict:
             return data.to_dict()
         return data
+
+    def get_calendar(self, proxy=None) -> dict:
+        self._quote.proxy = proxy or self.proxy
+        return self._quote.calendar
 
     def get_major_holders(self, proxy=None, as_dict=False):
         self._holders.proxy = proxy or self.proxy
@@ -1736,6 +1752,30 @@ class TickerBase:
             if as_dict:
                 return data.to_dict()
             return data
+    
+    def get_insider_purchases(self, proxy=None, as_dict=False):
+        self._holders.proxy = proxy or self.proxy
+        data = self._holders.insider_purchases
+        if data is not None:
+            if as_dict:
+                return data.to_dict()
+            return data
+
+    def get_insider_transactions(self, proxy=None, as_dict=False):
+        self._holders.proxy = proxy or self.proxy
+        data = self._holders.insider_transactions
+        if data is not None:
+            if as_dict:
+                return data.to_dict()
+            return data
+
+    def get_insider_roster_holders(self, proxy=None, as_dict=False):
+        self._holders.proxy = proxy or self.proxy
+        data = self._holders.insider_roster
+        if data is not None:
+            if as_dict:
+                return data.to_dict()
+            return data
 
     def get_info(self, proxy=None) -> dict:
         self._quote.proxy = proxy or self.proxy
@@ -1755,13 +1795,6 @@ class TickerBase:
     def get_sustainability(self, proxy=None, as_dict=False):
         self._quote.proxy = proxy or self.proxy
         data = self._quote.sustainability
-        if as_dict:
-            return data.to_dict()
-        return data
-
-    def get_recommendations_summary(self, proxy=None, as_dict=False):
-        self._quote.proxy = proxy or self.proxy
-        data = self._quote.recommendations
         if as_dict:
             return data.to_dict()
         return data
@@ -1968,11 +2001,9 @@ class TickerBase:
         if start is not None:
             start_ts = utils._parse_user_dt(start, tz)
             start = pd.Timestamp.fromtimestamp(start_ts).tz_localize("UTC").tz_convert(tz)
-            start_d = start.date()
         if end is not None:
             end_ts = utils._parse_user_dt(end, tz)
             end = pd.Timestamp.fromtimestamp(end_ts).tz_localize("UTC").tz_convert(tz)
-            end_d = end.date()
         if end is None:
             end = dt_now
         if start is None:
@@ -1994,7 +2025,7 @@ class TickerBase:
             return None
         try:
             fail = json_data["finance"]["error"]["code"] == "Bad Request"
-        except KeyError as e:
+        except KeyError:
             fail = False
         if fail:
             logger.error(f"{self.ticker}: Yahoo web request for share count failed")
@@ -2096,7 +2127,7 @@ class TickerBase:
                                    "the issue. Thank you for your patience.")
 
             try:
-                data = pd.read_html(data)[0]
+                data = pd.read_html(StringIO(data))[0]
             except ValueError:
                 if page_offset == 0:
                     # Should not fail on first page
@@ -2129,7 +2160,7 @@ class TickerBase:
 
         # Convert types
         for cn in ["EPS Estimate", "Reported EPS", "Surprise(%)"]:
-            dates.loc[dates[cn] == '-', cn] = "NaN"
+            dates.loc[dates[cn] == '-', cn] = float("nan")
             dates[cn] = dates[cn].astype(float)
 
         # Convert % to range 0->1: